requirements-test.txt
# Testing Dependencies for Trading Platform Backend

# Core Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1

# Code Coverage
pytest-cov==4.1.0

# HTML Reporting
pytest-html==4.1.1

# Async HTTP Client (for load testing)
aiohttp==3.9.1

# Sync HTTP Client (alternative to aiohttp)
httpx==0.25.1

# Database Testing
sqlalchemy==2.0.23

# FastAPI Testing
fastapi==0.104.1

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Security
python-jose==3.3.0
cryptography==41.0.7

# Rate Limiting
slowapi==0.1.9

# Redis (if using cache in tests)
redis==5.0.1

# Environment Variables
python-dotenv==1.0.0

# Additional Testing Utilities
pytest-timeout==2.1.0          # Timeout for long-running tests
pytest-xdist==3.5.0            # Parallel test execution
pytest-mock==3.12.0            # Mocking utilities
freezegun==1.4.0               # Time mocking

# Database Drivers
pymysql==1.1.0
# sqlite3 is built-in to Python

# JSON Schema Validation
jsonschema==4.20.0

# YAML (if needed for fixtures)
pyyaml==6.0.1
